requirements.txt
torch
torch_geometric
numpy
scipy
tqdm
matplotlib
pygsp
sortedcontainers
networkx
igraph
torch_sparse
torch_scatter
leidenalg
tensorboard
ogb